+++ src/index.js
@@ -1,4 +1,3 @@
 export * from './floating-select/index';
-export * from './floating-select-portal/index';
 export * from './floating-select-infinite/index';
-export styles from './floating-select/styles';
+export styles from './base/styles';