+++ src_docs/app.component.jsx
@@ -1,12 +1,12 @@
 import React from 'react';
-import { BrowserRouter as Router, Route } from 'react-router-dom';
-import { hot } from 'react-hot-loader';
+import { HashRouter as Router, Route } from 'react-router-dom';
+import { hot } from 'react-hot-loader/root';
 import ExampleContainer from './containers/example.container';
 
 import './app.component.scss';
 import './images/favicon.ico';
 
-export default @hot(module) class App extends React.PureComponent {
+class App extends React.PureComponent {
   render() {
     return (
       <Router>
@@ -15,3 +15,5 @@ export default @hot(module) class App extends React.PureComponent {
     );
   }
 }
+
+export default hot(App);